Trait r3_core::bind::RuntimeBinder

pub trait RuntimeBinder: Send + Copy + 'static {
    type Target<'call>;

    unsafe fn materialize<'call>(self) -> Self::Target<'call>;
}

Unstable. The runtime representation of Binder.

This trait signifies the following properties regarding an implementing type:

self can be “materialized” as Self::Target<'call> at runtime.

'call represets the duration during which Self::Target is used. If Self::Target<'call> is a reference, its lifetime parameter may be bound to 'call. Some binder types don’t require this.

Stability

This trait is unstable.

Required Associated Types

type Target<'call>

The materialized form.

Required Methods

unsafe fn materialize<'call>(self) -> Self::Target<'call>

Construct a target object at runtime, using the intermediate product constructed by Binder::into_runtime_binder.

Safety

intermediate must have been constructed by <Self as Binder>::into_runtime_binder.

The caller must uphold that Self::Target is safe to exist. (The configuration system is reponsible for enforcing this property.)