Struct r3_core::kernel::semaphore::SemaphoreRef

#[repr(transparent)]
pub struct SemaphoreRef<'a, System: KernelSemaphore>(_, _);

Represents a single borrowed semaphore in a system.

This type is ABI-compatible with System::RawSemaphoreId. It’s logically equivalent to &'a Semaphore but instead stores RawSemaphoreId directly.

See Semaphore for the owned counterpart and the description of this kernel object. See SemaphoreMethods for the operations provided by this handle type.

Implementations

impl<System: KernelSemaphore> SemaphoreRef<'static, System>

pub const fn define() -> SemaphoreDefiner<System>

Construct a SemaphoreDefiner to define a semaphore in a configuration function.