Type Definition r3_core::kernel::semaphore::StaticSemaphore

pub type StaticSemaphore<System> = SemaphoreRef<'static, System>;

A static handle type: SemaphoreRef<'static, System>

This type is ABI-compatible with System::RawSemaphoreId. It’s logically equivalent to &'static Semaphore but instead stores RawSemaphoreId directly.

See Semaphore for the owned counterpart and the description of this kernel object. See SemaphoreMethods for the operations provided by this handle type.

Implementations

impl<System: KernelSemaphore> StaticSemaphore<System>

pub const fn define() -> SemaphoreDefiner<System>

Construct a SemaphoreDefiner to define a semaphore in a configuration function.