Struct r3_core::kernel::timer::TimerRef

#[repr(transparent)]
pub struct TimerRef<'a, System: KernelTimer>(_, _);

Represents a single borrowed timer in a system.

This type is ABI-compatible with System::RawTimerId. It’s logically equivalent to &'a Timer but instead stores RawTimerId directly.

See Timer for the owned counterpart and the description of this kernel object. See TimerMethods for the operations provided by this handle type.

Implementations

impl<System: KernelTimer> TimerRef<'static, System>

pub const fn define() -> TimerDefiner<System>

Construct a TimerDefiner to define a timer in a configuration function.