Struct r3_core::time::Duration

#[repr(transparent)]
pub struct Duration { /* private fields */ }

Represents a signed time span used by the API surface of R3-OS.

Duration is backed by i32 and can represent the range [-35′47.483648″, +35′47.483647″] with microsecond precision.

Implementations

impl Duration

pub const ZERO: Self = _

An empty interval.

pub const MAX: Self = _

The large representable positive time span (+35′47.483647″).

pub const MIN: Self = _

The large representable negative time span (-35′47.483648″).

pub const fn from_micros(micros: i32) -> Self

Construct a new Duration from the specified number of microseconds.

pub const fn from_millis(millis: i32) -> Self

Construct a new Duration from the specified number of milliseconds.

Pancis if millis overflows the representable range of Duration.

pub const fn from_secs(secs: i32) -> Self

Construct a new Duration from the specified number of seconds.

Pancis if secs overflows the representable range of Duration.

pub const fn as_micros(self) -> i32

Get the total number of whole microseconds contained by this Duration.

pub const fn as_millis(self) -> i32

Get the total number of whole milliseconds contained by this Duration.

pub const fn as_secs(self) -> i32

Get the total number of whole seconds contained by this Duration.

pub const fn as_secs_f64(self) -> f64

Get the total number of seconds contained by this Duration as f64.

Examples

use r3_core::time::Duration;

let dur = Duration::from_micros(1_201_250_000);
assert_eq!(dur.as_secs_f64(), 1201.25);

pub const fn as_secs_f32(self) -> f32

Get the total number of seconds contained by this Duration as f32.

Examples

use r3_core::time::Duration;

let dur = Duration::from_micros(1_201_250_000);
assert_eq!(dur.as_secs_f32(), 1201.25);

pub const fn is_positive(self) -> bool

Return true if and only if self is positive.

pub const fn is_negative(self) -> bool

Return true if and only if self is negative.

pub const fn checked_mul(self, other: i32) -> Option<Self>

Multiply self by the specified value, returning None if the result overflows.

pub const fn checked_div(self, other: i32) -> Option<Self>

Divide self by the specified value, returning None if the result overflows or other is zero.

pub const fn checked_abs(self) -> Option<Self>

Calculate the absolute value of self, returning None if self == MIN.

pub const fn checked_add(self, other: Self) -> Option<Self>

Add the specified value to self, returning None if the result overflows.

pub const fn checked_sub(self, other: Self) -> Option<Self>

Subtract the specified value from self, returning None if the result overflows.