Module r3_core::utils

Utility

This module is exempt from the API stability guarantee unless specified otherwise. It’s exposed mostly because it’s needed by macros.

Re-exports

Zeroable, ZeroableInOption: Re-exported from bytemuck ^1. Zeroable is a marker trait used in HunkDefiner::zeroed and suchlike. It can be derived for struct types. These re-exports are subject to the application-side API stability guarantee.

Modules

for_times

mem

Structs

AlignedStorage

Untyped storage of the specified size and alignment. This is analogous to C++’s std::aligned_storage_t.

AllocError

The AllocError error indicates an allocation failure that may be due to resource exhaustion or to something wrong when combining the given input arguments with this allocator.

ConstAllocator

Compile-time allocator.

Frozen

Erases interior mutability by preventing reference forming.

RawCell

Like UnsafeCell, but implements Sync.

Traits

Allocator

const fn-compatible core::alloc::Allocator.

Init

Trait for types having a constant default value. This is essentially a constant version of Default.

Zeroable

Trait for types that can be safely created with zeroed.

ZeroableInOption

Trait for types which are Zeroable when wrapped in Option.

Type Definitions

PhantomInvariant

A phantom type that is invariant over T.

Derive Macros

Zeroable

Derive the Zeroable trait for a struct